Phase 8: Tier 3 — Iterative/Eigenvalue Domains (45 domains) — 78/78 gauntlet
Wire 45 iterative/eigenvalue physics domains into the STARK proof pipeline:

Quantum Mechanics (5): TISE, TDSE, Scattering, Semiclassical, Path Integrals
Quantum Many-Body (13): DMRG, Quantum Spin, Strongly Correlated, Topological,
  MBL, Lattice Gauge, Open Quantum, Non-equilibrium QM, Kondo/Impurity,
  Bosonic, Fermionic, Nuclear Many-Body, Ultracold Atoms
Electronic Structure (7): DFT, Beyond-DFT, Tight Binding, Excited States,
  Response Properties, Relativistic, Embedding
Solid State (8): Phonons, Band Structure, Classical Magnetism,
  Superconductivity, Disordered, Surfaces, Defects, Ferroelectrics
Nuclear & Particle (6): Nuclear Structure, Nuclear Reactions, Nuclear Astro,
  Lattice QCD, Perturbative QFT, Beyond Standard Model
Chemical Physics (3): PES, Reaction Rate, Catalysis
Quantum Information (3): Quantum Circuit, QEC, VQE

Infrastructure:
- SCF + eigenvalue core trace adapters
- 45 domain trace adapters across 10 directories
- 7 Lean proof files (71 fixed-point conservation theorems)
- TPC generator (scripts/tpc/generate_phase8.py)
- Phase 8 gauntlet (78 tests across 13 validation layers)
- Fix: hybrid.py SVD dtype mismatch (complex128 @ float64)

+++ tensornet/quantum/hybrid.py
@@ -333,7 +333,7 @@ class TNQuantumSimulator:
         Vh = Vh[:chi_new, :]
 
         # Absorb singular values into left tensor
-        US = U @ torch.diag(S)
+        US = U @ torch.diag(S.to(U.dtype))
 
         # Reshape back to MPS tensors
         self.mps[q_left] = US.reshape(chi_l, 2, chi_new)
@@ -416,7 +416,7 @@ class TNQuantumSimulator:
         S = S[:chi_new]
         Vh = Vh[:chi_new, :]
 
-        US = U @ torch.diag(S)
+        US = U @ torch.diag(S.to(U.dtype))
 
         self.mps[q_left] = US.reshape(chi_l, 2, chi_new)
         self.mps[q_right] = Vh.reshape(chi_new, 2, chi_r)
